Added discussion bot capabilities to phpBB plugin (sorry guys, the rest of the plugins are up to you).  Fixed a few issues with the discussion and search plugin api.
Decided against cbparser so removed it.  Added NBBC (BSD Open Source License) to parse html to bbcode and created my own simple parser for bbcode to html.  

Fixed up vbulletin authentication plugin details and added an autoinstaller of it when the vB dual login plugin is enabled via the config page.  

Added a flag to JFusionFunction::lookupUser to use the Joomla ID as a reference (needed for discussion bot).