Fixed a couple bugs in the user sync handling where the link to conflict was never displayed, the error counter for each plugin was missing, and where fatal errors occured when resolving conflicts as the error data is no longer stored in the database.
Also fixed a bug in the user plugin where Joomla's userinfo was passed into the deleteUser functions of the plugins rather than the plugin's userinfo.  Commented out the code that deletes the user from the master after a Joomla user is deleted as the master should take care of user management.